Don't remove files when $ENV{VERBOSE}

+++ t/C-COMPILED-CPANEL/template.pl
@@ -71,7 +71,7 @@ foreach my $optimization (@optimizations) {
         ok( -e $c_file && !-z _, "$c_file is generated ($optimization)" );
 
         if ( -z $c_file ) {
-            unlink $c_file;
+            unlink $c_file unless $ENV{BC_DEVELOPING};
             skip( "Can't test further due to failure to create a c file.", 9 );
         }
 
@@ -87,7 +87,7 @@ foreach my $optimization (@optimizations) {
         ok( -x $bin_file, "$bin_file is compiled and ready to run." );
 
         if ( !-x $bin_file ) {
-            unlink $c_file, $bin_file;
+            unlink $c_file, $bin_file unless $ENV{BC_DEVELOPING};
             skip( "Can't test further due to failure to create a binary file.", 8 );
         }
